generator/requirements.txt
python-dotenv==1.0.0
PyYAML==6.0
paramiko==3.1.0
